enemies drop scrap on death (#29)
- delete EntityType/EntityTypeComponent
- add some tests around moving on to items
- fov updates after enemy turn
- using badly name bags/ directory for data for the actors

+++ src/imgui/helpers.hpp
@@ -12,11 +12,11 @@ struct WomboComboIn
 {
   std::string label;
 
-  int current_index = 0;
+  size_t current_index = 0;
   const std::vector<std::string>& items;
 
   WomboComboIn(const std::vector<std::string>& i)
-    : items(i){};
+    : items(i) {};
 };
 struct WomboComboOut
 {
